Issue #11: Reorganized views into folder module.
